update notifications controllers

+++ src/controllers/notificationController.ts
@@ -1,6 +1,7 @@
 import DataCache from '../utilities/dataCache'
 import NotificationModel, { INotification, TNotificationType } from '../dataSource/models/notificationModel'
 import DataRequest, { IListOutput, IPgeInfo } from '../utilities/dataQuery'
+import accountController from './accountController'
 // import Config from '../utilities/config'
 
 // const env = Config.getEnv()
@@ -17,8 +18,14 @@ class NotificationController {
     public async getNotification(accountId:string, notifId:string):Promise<INotification|null> {
         if (!accountId || !notifId) return null
         // check that the account existed
+        const account = await accountController.getAccount({_id: accountId})
+        if (!account) throw({code: 404})
+
         // check if the ntif is really under this account
-        return await this.cachedData.getItem<INotification>(notifId)
+        const notif = await this.cachedData.getItem<INotification>(notifId)
+        if (notif?.accountId !== account._id) throw({code: 404})
+
+        return notif
     }
 
     public async getNotificationsByPage(query:{accountId?:string, seen?:boolean} = {}, pageInfo: IPgeInfo):Promise<IListOutput<INotification>> {
@@ -36,7 +43,7 @@ class NotificationController {
         return result
     }
 
-    public async updateNotification(id:string, accountId:string, type:TNotificationType, title:string, message:string, link?:string, seen?:boolean):Promise<INotification | null> {
+    public async updateNotification(accountId:string, notifId:string, type:TNotificationType, title:string, message:string, link?:string, seen?:boolean):Promise<INotification | null> {
         const doc:INotification = {}
 
         if (accountId) doc.accountId = accountId
@@ -46,7 +53,15 @@ class NotificationController {
         if (link) doc.link = link
         if (seen) doc.seen = seen
 
-        const result = await this.cachedData.updateItem<INotification>(id, doc)
+        // check that the account existed
+        const account = await accountController.getAccount({_id: accountId})
+        if (!account) throw({code: 404})
+
+        // check if the ntif is really under this account
+        const notif = await this.cachedData.getItem<INotification>(notifId)
+        if (notif?.accountId !== account._id) throw({code: 404})
+
+        const result = await this.cachedData.updateItem<INotification>(notifId, doc)
 
         return result
     }
@@ -55,7 +70,12 @@ class NotificationController {
 
         if (!accountId || !notifId) return null
         // check that the account existed
+        const account = await accountController.getAccount({_id: accountId})
+        if (!account) throw({code: 404})
+
         // check if the ntif is really under this account
+        const notif = await this.cachedData.getItem<INotification>(notifId)
+        if (notif?.accountId !== account._id) throw({code: 404})
 
         const result = await this.cachedData.deleteItem<INotification>(notifId)
